pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pypenguin"  # Ensure this is unique on PyPI
version = "0.1.0"
authors = [{ name = "Fritzforcode" }]
description = "Python tool for creating, editing, and inspecting Penguinmod (.pmp) files"
readme = "README.md"
requires-python = ">=3.7"

dependencies = [
    "pillow", 
    "jsoncomment",
    "requests",
]

[tool.setuptools]
packages = ["pypenguin"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"pypenguin" = ["assets/*"]  # If non-Python files exist

[project.urls]
Homepage = "https://github.com/your-github-username/PyPenguin"
Repository = "https://github.com/your-github-username/PyPenguin"
Issues = "https://github.com/your-github-username/PyPenguin/issues"
